+++ nodejs/aws-infra/clusterMod/cluster2.ts
@@ -15,11 +15,14 @@
 import * as aws from "@pulumi/aws";
 import * as pulumi from "@pulumi/pulumi";
 
-import { Network } from "./network";
-import { Overwrite, sha1hash } from "./utils";
+import { ClusterFileSystem, ClusterFileSystemArgs } from "./clusterFileSystem";
+import { ClusterLoadBalancer, ClusterLoadBalancerArgs } from "./clusterLoadBalancer";
+
+import { Network } from "./../network";
+import { Overwrite, sha1hash } from "./../utils";
 
 /**
- * Arguments bag for creating infrastrcture for a new Cluster.
+ * Arguments bag for creating infrastructure for a new Cluster.
  */
 export interface ClusterArgs2 {
     /**
@@ -32,29 +35,6 @@ export interface ClusterArgs2 {
      * created.
      */
     instanceSecurityGroup?: aws.ec2.SecurityGroup;
-}
-
-/**
- * Arguments for creating a file system for a cluster.
- */
-export interface FileSystemArgs2 {
-    /**
-     * The security group to use for the file system.  If not provided, a default one that allows
-     * ingress for the cluster's VPC from port 2049 will be created.
-     */
-    securityGroup?: aws.ec2.SecurityGroup;
-
-    /**
-     * The subnets to mount the file system against.  If not provided, file system will be mounted
-     * for every subnet in the cluster's network.
-     */
-    subnetIds?: pulumi.Input<string>[];
-
-    /**
-     * Path to mount file system at when a cluster is connected to an autoscaling group.  If not
-     * provided, the default mountPath will be "/mnt/efs"
-     */
-    mountPath?: pulumi.Input<string>;
 }
 
 export interface AutoScalingGroupArgs2 {
@@ -193,41 +173,6 @@ export type LaunchConfigurationArgs = Overwrite<aws.ec2.LaunchConfigurationArgs,
 }>;
 
 /**
- * ContainerPort represents the information about how to expose a container port on a [Service].
-*/
-export interface ClusterLoadBalancerArgs {
-    /**
-     * The incoming port where the service exposes the endpoint.
-    */
-    port: number;
-    /**
-     * The target port on the backing container.  Defaults to the value of [port].
-    */
-    targetPort?: number;
-    /**
-     * Whether the port should be exposed externally.  Defaults to `false`.
-    */
-    external?: boolean;
-    /**
-     * The protocol to use for exposing the service:
-     *  `tcp`: Expose TCP externaly and to the container.  Will create a network load balancer.
-     *  `http`: Expose HTTP externally and to the container.  Will create application load balancer.
-     *  `https`: Expose HTTPS externally and HTTP to the container.   Will create application load balancer.
-     *
-     * Defaults to 'tcp' if unspecified.
-     */
-    protocol?: "tcp" | "http" | "https";
-
-    /**
-     * The ARN of the default SSL server certificate. Exactly one certificate is required if the
-     * protocol is [https]. For adding additional SSL certificates, see the
-     * [`aws_lb_listener_certificate`
-     * resource](https://www.terraform.io/docs/providers/aws/r/lb_listener_certificate.html).
-     */
-    certificateArn?: string;
-}
-
-/**
  * A Cluster is a general purpose ECS cluster configured to run in a provided Network.
  */
 export class Cluster2 extends pulumi.ComponentResource {
@@ -299,37 +244,8 @@ export class Cluster2 extends pulumi.ComponentResource {
         });
     }
 
-    public createFileSystem(name: string, args: FileSystemArgs2): aws.efs.FileSystem {
-        // If requested, add EFS file system and mount targets in each subnet.
-        const parentOpts = { parent: this };
-
-        const fileSystem = new aws.efs.FileSystem(name, {}, parentOpts);
-        const efsSecurityGroupName = `${name}-fs`;
-        const efsSecurityGroup = args.securityGroup || new aws.ec2.SecurityGroup(efsSecurityGroupName, {
-            vpcId: this.network.vpcId,
-            ingress: [
-                // Allow NFS traffic from the instance security group
-                {
-                    securityGroups: [ this.instanceSecurityGroup.id ],
-                    protocol: "TCP",
-                    fromPort: 2049,
-                    toPort: 2049,
-                },
-            ],
-            tags: { Name: efsSecurityGroupName },
-        }, parentOpts);
-
-        const subnetIds = args.subnetIds || this.network.subnetIds;
-        for (let i = 0; i < subnetIds.length; i++) {
-            const subnetId = subnetIds[i];
-            const mountTarget = new aws.efs.MountTarget(`${name}-${i}`, {
-                fileSystemId: fileSystem.id,
-                subnetId: subnetId,
-                securityGroups: [ efsSecurityGroup.id ],
-            }, parentOpts);
-        }
-
-        return fileSystem;
+    public createFileSystem(name: string, args: ClusterFileSystemArgs): ClusterFileSystem {
+        return new ClusterFileSystem(name, this, args);
     }
 
     /**
@@ -384,110 +300,6 @@ export class Cluster2 extends pulumi.ComponentResource {
      */
     public createLoadBalancer(name: string, args: ClusterLoadBalancerArgs): ClusterLoadBalancer {
         return new ClusterLoadBalancer(name, this, args);
-    }
-}
-
-export class ClusterLoadBalancer extends pulumi.ComponentResource {
-    public readonly cluster: Cluster2;
-    public readonly loadBalancer: aws.elasticloadbalancingv2.LoadBalancer;
-    public readonly targetGroup: aws.elasticloadbalancingv2.TargetGroup;
-    public readonly listener: aws.elasticloadbalancingv2.Listener;
-
-    constructor(name: string, cluster: Cluster2, args: ClusterLoadBalancerArgs) {
-        super("aws.infra.ClusterLoadBalancer", name, args);
-        this.cluster = cluster;
-
-        // Load balancers need *very* short names, so we unfortunately have to hash here.
-        //
-        // Note: Technically, we can only support one LB per service, so only the service name is needed here, but we
-        // anticipate this will not always be the case, so we include a set of values which must be unique.
-        const longName = `${name}-${args.port}`;
-        const shortName = sha1hash(`${longName}`);
-
-        // Create an internal load balancer if requested.
-        const internal = cluster.network.usePrivateSubnets && !args.external;
-
-        // See what kind of load balancer to create (application L7 for HTTP(S) traffic, or network L4 otherwise).
-        // Also ensure that we have an SSL certificate for termination at the LB, if that was requested.
-        const { listenerProtocol, targetProtocol, useAppLoadBalancer, certificateArn } =
-            computeLoadBalancerInfo(args);
-
-        const parentOpts = { parent: this };
-        this.loadBalancer = new aws.elasticloadbalancingv2.LoadBalancer(shortName, {
-            loadBalancerType: useAppLoadBalancer ? "application" : "network",
-            subnets: cluster.network.publicSubnetIds,
-            internal: internal,
-            // If this is an application LB, we need to associate it with the ECS cluster's security
-            // group, so that traffic on any ports can reach it.  Otherwise, leave blank, and
-            // default to the VPC's group.
-            securityGroups: useAppLoadBalancer ? [ cluster.instanceSecurityGroup.id ] : undefined,
-            tags: { Name: longName },
-        }, parentOpts);
-
-        // Create the target group for the new container/port pair.
-        this.targetGroup = new aws.elasticloadbalancingv2.TargetGroup(shortName, {
-            port: args.targetPort || args.port,
-            protocol: targetProtocol,
-            vpcId: cluster.network.vpcId,
-            deregistrationDelay: 180, // 3 minutes
-            tags: { Name: longName },
-            targetType: "ip",
-        }, parentOpts);
-
-        // Listen on the requested port on the LB and forward to the target.
-        this.listener = new aws.elasticloadbalancingv2.Listener(longName, {
-            loadBalancerArn: this.loadBalancer.arn,
-            protocol: listenerProtocol,
-            certificateArn: certificateArn,
-            port: args.port,
-            defaultAction: {
-                type: "forward",
-                targetGroupArn: this.targetGroup.arn,
-            },
-            // If SSL is used, we automatically insert the recommended ELB security policy from
-            // http://docs.aws.amazon.com/elasticloadbalancing/latest/application/create-https-listener.html.
-            sslPolicy: certificateArn ? "ELBSecurityPolicy-2016-08" : undefined,
-        }, parentOpts);
-
-        this.registerOutputs({
-            loadBalancer: this.loadBalancer,
-            targetGroup: this.targetGroup,
-            listener: this.listener,
-        });
-    }
-}
-
-function computeLoadBalancerInfo(args: ClusterLoadBalancerArgs) {
-    switch (args.protocol || "tcp") {
-        case "https":
-            if (!args.certificateArn) {
-                throw new Error("Cannot create Service for HTTPS trafic. No ACM certificate ARN configured.");
-            }
-
-            return {
-                listenerProtocol: "HTTPS",
-                // Set the target protocol to HTTP, so that the ELB terminates the SSL traffic.
-                // IDEA: eventually we should let users choose where the SSL termination occurs.
-                targetProtocol: "HTTP",
-                useAppLoadBalancer: true,
-                certificateArn: args.certificateArn,
-            };
-        case "http":
-            return {
-                listenerProtocol: "HTTP",
-                targetProtocol: "HTTP",
-                useAppLoadBalancer: true,
-                certificateArn: undefined,
-            };
-        case "tcp":
-            return {
-                listenerProtocol: "TCP",
-                targetProtocol: "TCP",
-                useAppLoadBalancer: false,
-                certificateArn: undefined,
-            };
-        default:
-            throw new Error(`Unrecognized Service protocol: ${args.protocol}`);
     }
 }
