+++ nodejs/aws-infra/clusterMod/index.ts
@@ -12,6 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-export * from "./cluster";
-export * from "./network";
-export * from "./clusterMod";
+export * from "./cluster2";
+export * from "./clusterFileSystem";
+export * from "./clusterLoadBalancer";
